modifica dei file per serie di fibonacci, ottimizzazione C

+++ E11/serie.cpp
@@ -1,9 +1,9 @@
 #include <iostream>
 
-extern "C" {
+extern "C" { 
 
-double fibonacci(int n) {
-    double fib = 1, fob = 1, pp;
+long double fibonacci(int n) {
+    long double fib = 1, fob = 1, pp; // fib è F_n, fob è F_n-1 
     if (n > 2) {
         for (int i = 2; i != n; i++) {
             pp = fib;
